can delete images associated with deleted products

+++ express-basics/controllers/admin.js
@@ -1,4 +1,5 @@
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
 const {
   validationResult
@@ -138,6 +139,7 @@ exports.postEditProduct = (req, res, next) => {
         product.price = updatedPrice;
         product.description = updatedDescription;
         if(image){
+          fileHelper.deleteFile(product.imageUrl);
           product.imageUrl = image.path;
         }
       return product.save()
@@ -171,12 +173,19 @@ exports.getProducts = (req, res, next) => {
     });
 }
 
-exports.postDeleteProduct = (req, res) => {
+exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({
-      _id: prodId,
-      userId: req.user._id
-    })
+  Product.findById(prodId)
+  .then(product => {
+    if(!product){
+      next(new Error('Product not found.'))
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({
+        _id: prodId,
+        userId: req.user._id
+      })
+  })
     .then(() => res.redirect('/admin/products'))
     .catch(err => {
       const error = new Error(err);
